+++ src/memiter.c
@@ -52,7 +52,7 @@ static void memiter_skip_space(struct memiter *it)
  */
 bool memiter_iseq(const struct memiter *it, const char *str)
 {
-	size_t it_len = it->limit - it->next;
+	size_t it_len = (uintptr_t)it->limit - (uintptr_t)it->next;
 	size_t len = strnlen_s(str, it_len + 1);
 
 	if (len != it_len) {
@@ -177,5 +177,5 @@ const void *memiter_base(const struct memiter *it)
  */
 size_t memiter_size(const struct memiter *it)
 {
-	return it->limit - it->next;
+	return (size_t)(it->limit - it->next);
 }
